pyproject.toml
[project]
name = "policy-dq"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "pytest>=9.0.2",
    "pyyaml>=6.0.3",
    "requests>=2.33.1",
    "typer>=0.24.1",
    "uvicorn>=0.44.0",
]

[dependency-groups]
dev = [
    "mypy>=1.20.0",
    "ruff>=0.15.9",
]

[project.scripts]
policy-dq = "policy_dq.cli:cli"
policy-dq-api = "uvicorn:main"
